requirements.txt
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
contourpy==1.2.1
cryptography==42.0.5
cycler==0.12.1
filelock==3.13.3
fonttools==4.50.0
fsspec==2024.3.1
gcloud==0.18.3
gitdb==4.0.11
GitPython==3.1.43
googleapis-common-protos==1.63.0
httplib2==0.22.0
idna==3.6
importlib_resources==6.4.0
Jinja2==3.1.3
jwcrypto==1.5.6
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.3
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
oauth2client==4.1.3
opencv-contrib-python==4.9.0.80
packaging==24.0
pandas==2.2.1
pillow==10.3.0
protobuf==4.25.3
psutil==5.9.8
py-cpuinfo==9.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodome==3.20.0
pyparsing==3.1.2
Pyrebase4==4.7.1
python-dateutil==2.9.0.post0
python-jwt==4.1.0
pytz==2024.1
PyYAML==6.0.1
requests==2.29.0
requests-toolbelt==0.10.1
rsa==4.9
scipy==1.13.0
seaborn==0.13.2
six==1.16.0
smmap==5.0.1
sympy==1.12
thop==0.1.1.post2209072238
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
tqdm==4.66.2
typing_extensions==4.10.0
tzdata==2024.1
ultralytics==8.1.42
urllib3==1.26.18
zipp==3.18.1
